requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
openai==1.3.0
streamlit==1.28.0
python-dotenv==1.0.0
pydantic==2.4.2
PyPDF2==3.0.1
spacy==3.7.2
sqlalchemy==2.0.23
pytest==7.4.3
python-multipart==0.0.6
aiosqlite==0.19.0
langchain==0.0.335
